[Feat] SearchPage Feature And Linked DatePicker
