+++ package.json
@@ -11,6 +11,7 @@
     "moment": "^2.29.4",
     "persist": "^0.2.7",
     "react": "^18.2.0",
+    "react-datepicker": "^4.11.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.8.0",
     "react-persist": "^1.0.2",
