+++ src/Routes/Router.js
@@ -10,6 +10,7 @@ import PostPage from "../Pages/Post/PostPage/PostPage";
 import MyPostsContainer from "../Pages/MyPost/MyPostsContainer";
 import MyScrapContainer from "../Pages/MyScrap/MyScrapContainer";
 import ScrapInterviewContainer from "../Pages/ScrapInterview/ScrapInterviewContainer";
+import SearchPage from "../Pages/Search/SearchPage";
 import GoogleSocialLogin from "../Pages/SocialLogin/GoogleSocialLogin";
 import NaverSocialLogin from "../Pages/SocialLogin/NaverSocialLogin";
 import KakaoSocialLogin from "../Pages/SocialLogin/KakaoSocialLogin";
@@ -26,6 +27,7 @@ const Router = () => {
       <Route path="/scrap" element={<MyScrapContainer />} />
       <Route path="/interview/:interview_id" element={<InterviewConatiner />} />
       <Route path="/scraps/:scrap_id" element={<ScrapInterviewContainer />} />
+      <Route path="/search" element={<SearchPage />} />
       <Route path="/*" element={<PageNotFound />} />
       <Route path="/oauth2/google/redirect" element={<GoogleSocialLogin/>}/>
       <Route path="/oauth2/naver/redirect" element={<NaverSocialLogin/>}/>
